generating from cache

+++ models/daily_stats.go
@@ -4,8 +4,8 @@ import "time"
 
 type DailyStatsByCity struct {
 	AveragePrice          int
-	AveragePricePerSquare float32
-	AverageFootage        float32
+	AveragePricePerSquare float64
+	AverageFootage        float64
 	AdsCount              int
 	Date                  time.Time
 	City                  string
